Reduce number of glEnable and glDisable calls
There is no need for postDraw to disable blending on every postDraw().
It is enough if Shape will make sure that it has the correct blending
setting in preDraw() method.

Since 99% of Shapes are drawn with blending enabled, this will
substantially reduce number of unnecessary JNI calls, and (hopefully)
improve overall performance.