Reduce number of glEnable and glDisable calls
There is no reason for postDraw() to disable blending on every call.
It will be enough if Shape makes sure that it has the correct blending
setting in preDraw() method.

Since 99% of Shapes are drawn with blending enabled, applying this
patch substantially reduce number of redundand JNI calls, and (hopefully)
improve overall performance.