Changed how ScrollDetector processes touches.
Changed ScrollDetector so that it does not consume the move touch event
unless it triggers a scroll start and changed it so that up touch events
are not consumed unless a scroll was already in progress. This allows
entities to still receive touch events even if they are inside a
ScrollDetector touch area.